refactor(store-data): Colocate StorageProviderCard components
Moves the `StorageProviderCard` and its sub-components (`CardKeyFeature`, `CardSection`) into a dedicated directory for better organization and component colocation.

The sub-components have been renamed to be more descriptive (e.g., `StorageProviderCardKeyFeature`) to improve clarity and avoid potential naming conflicts.

Additionally, this commit fixes a layout issue by adjusting the margin on the card's action button.
