+++ apps/filecoin-site/src/app/store-data/components/StorageProviderCard/StorageProviderCardKeyFeature.tsx
@@ -2,11 +2,13 @@ import { CheckIcon } from '@phosphor-icons/react/dist/ssr'
 
 import { Icon } from '@filecoin-foundation/ui/Icon'
 
-export type CardKeyFeatureProps = {
+export type StorageProviderCardKeyFeatureProps = {
   feature: string
 }
 
-export function CardKeyFeature({ feature }: CardKeyFeatureProps) {
+export function StorageProviderCardKeyFeature({
+  feature,
+}: StorageProviderCardKeyFeatureProps) {
   return (
     <div className="flex items-center gap-2.5">
       <Icon component={CheckIcon} color="success" size={16} />
