+++ apps/filecoin-site/src/app/store-data/components/StorageProviderCard/StorageProviderCardSection.tsx
@@ -1,9 +1,12 @@
-type CardSectionProps = {
+type StorageProviderCardSectionProps = {
   title: string
   children: React.ReactNode
 }
 
-export function CardSection({ title, children }: CardSectionProps) {
+export function StorageProviderCardSection({
+  title,
+  children,
+}: StorageProviderCardSectionProps) {
   return (
     <div>
       <h4 className="mb-4 font-medium text-zinc-950">{title}</h4>
